Replace the race dumping program with a gecko code
The main reason for that change is that a gecko code was needed anyway
because the MemoryWatcher api can only read memory at the end of the
frame, but some variables like rot_vec2 are reinitialized before that.

Another benefit is that it can now work on non-UNIX systems and on real
Wii.

The unused rot_vec1 variable has been removed, speed1_soft_limit has
been added, and the variables have been reordered to make the dumping
code simpler so the version is bumped to 2.